refactor(web): move the session detail view into annotation-queues
structure:stats rule 6 reports every file under src/components/session as
used by annotation-queues only, so the folder moves there with git mv via
structure:move. The stories' relative path to .storybook/preview would have
had to change with the depth, which structure:move refuses (rule 15), so
they switch to the @/ alias first; two vi.mock() path strings and one now
inapplicable eslint-disable are fixed by hand.
