+++ web/src/features/annotation-queues/components/session/ModernSessionFilterDialogContent.stories.tsx
@@ -2,8 +2,8 @@ import { type ColumnDefinition, type FilterState } from "@langfuse/shared";
 import { type ComponentProps } from "react";
 import { fn } from "storybook/test";
 
-import preview from "../../../.storybook/preview";
-import { ModernSessionFilterDialogContent } from "@/src/components/session/ModernSessionFilterDialogContent";
+import preview from "@/.storybook/preview";
+import { ModernSessionFilterDialogContent } from "@/src/features/annotation-queues/components/session/ModernSessionFilterDialogContent";
 import { Dialog } from "@/src/components/ui/dialog";
 
 const filterColumns = [
