+++ web/src/features/annotation-queues/components/session/ModernSessionFilterDialogContent.tsx
