+++ web/src/features/annotation-queues/components/session/ModernSessionHeader.stories.tsx
@@ -1,9 +1,9 @@
 import { type ComponentProps, useState } from "react";
 import { expect, fn, userEvent, waitFor, within } from "storybook/test";
 
-import preview from "../../../.storybook/preview";
-import { ModernSessionHeader } from "@/src/components/session/ModernSessionHeader";
-import { sessionHeaderVisibilityStorageKey } from "@/src/components/session/sessionHeaderVisibility";
+import preview from "@/.storybook/preview";
+import { ModernSessionHeader } from "@/src/features/annotation-queues/components/session/ModernSessionHeader";
+import { sessionHeaderVisibilityStorageKey } from "@/src/features/annotation-queues/components/session/sessionHeaderVisibility";
 
 const scores = [
   {
