+++ web/src/features/annotation-queues/components/session/ModernSessionHeaderPill.tsx
