+++ web/src/features/annotation-queues/components/session/SessionObservationIO.tsx
@@ -9,7 +9,7 @@ import {
   type ViewMode,
 } from "@/src/features/traces";
 import { api, type RouterOutputs } from "@/src/utils/api";
-import { downloadJsonFile } from "@/src/components/session/actions/downloadSessionAsJson";
+import { downloadJsonFile } from "@/src/features/annotation-queues/components/session/actions/downloadSessionAsJson";
 import { showErrorToast } from "@/src/features/notifications/showErrorToast";
 import { usePostHogClientCapture } from "@/src/features/posthog-analytics/usePostHogClientCapture";
 import { compactNumberFormatter } from "@/src/utils/numbers";
