+++ web/src/features/annotation-queues/components/session/session-detail-presets.ts
