+++ web/src/features/annotation-queues/components/session/sessionHeaderVisibility.ts
