+++ web/src/features/annotation-queues/components/session/sessionIdleGap.ts
