+++ web/src/features/annotation-queues/components/session/sessionVisibleObservations.ts
