+++ web/src/features/annotation-queues/components/session/stableVirtualRowMeasurementState.ts
